Done refactoring for App.js
